Move equations to asym package
