+++ src/main/java/com/powsybl/openloadflow/ac/AcLoadFlowContext.java
@@ -6,7 +6,7 @@
  */
 package com.powsybl.openloadflow.ac;
 
-import com.powsybl.openloadflow.ac.equations.AsymmetricalAcEquationSystemCreator;
+import com.powsybl.openloadflow.ac.equations.asym.AsymmetricalAcEquationSystemCreator;
 import com.powsybl.openloadflow.lf.AbstractLoadFlowContext;
 import com.powsybl.openloadflow.ac.equations.AcEquationSystemCreator;
 import com.powsybl.openloadflow.ac.equations.AcEquationType;
@@ -37,7 +37,7 @@ public class AcLoadFlowContext extends AbstractLoadFlowContext<AcVariableType, A
     public EquationSystem<AcVariableType, AcEquationType> getEquationSystem() {
         if (equationSystem == null) {
             var creator = parameters.isAsymmetrical() ? new AsymmetricalAcEquationSystemCreator(network, parameters.getEquationSystemCreationParameters())
-                                               : new AcEquationSystemCreator(network, parameters.getEquationSystemCreationParameters());
+                                                      : new AcEquationSystemCreator(network, parameters.getEquationSystemCreationParameters());
             equationSystem = creator.create();
         }
         return equationSystem;
