+++ src/main/java/com/powsybl/openloadflow/ac/equations/asym/AbstractShuntFortescueCurrentEquationTerm.java
@@ -1,5 +1,6 @@
-package com.powsybl.openloadflow.ac.equations;
+package com.powsybl.openloadflow.ac.equations.asym;
 
+import com.powsybl.openloadflow.ac.equations.AcVariableType;
 import com.powsybl.openloadflow.equations.Variable;
 import com.powsybl.openloadflow.equations.VariableSet;
 import com.powsybl.openloadflow.network.LfBus;
