+++ src/main/java/com/powsybl/openloadflow/ac/equations/asym/AsymmetricalClosedBranchCoupledPowerEquationTerm.java
@@ -1,5 +1,6 @@
-package com.powsybl.openloadflow.ac.equations;
+package com.powsybl.openloadflow.ac.equations.asym;
 
+import com.powsybl.openloadflow.ac.equations.AcVariableType;
 import com.powsybl.openloadflow.equations.Variable;
 import com.powsybl.openloadflow.equations.VariableSet;
 import com.powsybl.openloadflow.network.LfBranch;
@@ -9,32 +10,33 @@ import com.powsybl.openloadflow.util.Fortescue;
 import java.util.Objects;
 
 /**
+ * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
  * @author Jean-Baptiste Heyberger <jbheyberger at gmail.com>
  */
-public class AsymmetricalClosedBranchCoupledCurrentEquationTerm extends AbstractAsymmetricalClosedBranchCoupledFlowEquationTerm {
+public class AsymmetricalClosedBranchCoupledPowerEquationTerm extends AbstractAsymmetricalClosedBranchCoupledFlowEquationTerm {
 
-    public AsymmetricalClosedBranchCoupledCurrentEquationTerm(LfBranch branch, LfBus bus1, LfBus bus2, VariableSet<AcVariableType> variableSet,
-                                                              boolean isRealPart, boolean isSide1, Fortescue.SequenceType sequenceType) {
+    public AsymmetricalClosedBranchCoupledPowerEquationTerm(LfBranch branch, LfBus bus1, LfBus bus2, VariableSet<AcVariableType> variableSet,
+                                                            boolean isRealPart, boolean isSide1, Fortescue.SequenceType sequenceType) {
         super(branch, bus1, bus2, variableSet, isRealPart, isSide1, sequenceType);
     }
 
-    public static double tx(int i, int j, int g, int h, AsymmetricalClosedBranchCoupledCurrentEquationTerm equationTerm) {
-        return GenericBranchCurrentTerm.iTx(i, j, g, h, equationTerm);
+    public static double tx(int i, int j, int g, int h, AsymmetricalClosedBranchCoupledPowerEquationTerm equationTerm) {
+        return GenericBranchPowerTerm.powerTx(i, j, g, h, equationTerm);
     }
 
-    public static double ty(int i, int j, int g, int h, AsymmetricalClosedBranchCoupledCurrentEquationTerm equationTerm) {
-        return GenericBranchCurrentTerm.iTy(i, j, g, h, equationTerm);
+    public static double ty(int i, int j, int g, int h, AsymmetricalClosedBranchCoupledPowerEquationTerm equationTerm) {
+        return GenericBranchPowerTerm.powerTy(i, j, g, h, equationTerm);
     }
 
-    public static double dtx(int i, int j, int g, int h, AsymmetricalClosedBranchCoupledCurrentEquationTerm equationTerm, Variable<AcVariableType> variable, int iDerivative) {
-        return GenericBranchCurrentTerm.idTx(i, j, g, h, equationTerm, variable, iDerivative);
+    public static double dtx(int i, int j, int g, int h, AsymmetricalClosedBranchCoupledPowerEquationTerm equationTerm, Variable<AcVariableType> variable, int iDerivative) {
+        return GenericBranchPowerTerm.powerdTx(i, j, g, h, equationTerm, variable, iDerivative);
     }
 
-    public static double dty(int i, int j, int g, int h, AsymmetricalClosedBranchCoupledCurrentEquationTerm equationTerm, Variable<AcVariableType> variable, int iDerivative) {
-        return GenericBranchCurrentTerm.idTy(i, j, g, h, equationTerm, variable, iDerivative);
+    public static double dty(int i, int j, int g, int h, AsymmetricalClosedBranchCoupledPowerEquationTerm equationTerm, Variable<AcVariableType> variable, int iDerivative) {
+        return GenericBranchPowerTerm.powerdTy(i, j, g, h, equationTerm, variable, iDerivative);
     }
 
-    public static double ixIyij(boolean isRealPart, boolean isSide1, int sequenceNum, AsymmetricalClosedBranchCoupledCurrentEquationTerm eqTerm) {
+    public static double pqij(boolean isRealPart, boolean isSide1, int sequenceNum, AsymmetricalClosedBranchCoupledPowerEquationTerm eqTerm) {
 
         int s1 = 1;
         int s2 = 2;
@@ -43,16 +45,16 @@ public class AsymmetricalClosedBranchCoupledCurrentEquationTerm extends Abstract
             s2 = 1;
         }
 
-        if (isRealPart) { // Ix
+        if (isRealPart) { // P
             return tx(s1, s1, sequenceNum, 0, eqTerm) + tx(s1, s1, sequenceNum, 1, eqTerm) + tx(s1, s1, sequenceNum, 2, eqTerm)
                     + tx(s1, s2, sequenceNum, 0, eqTerm) + tx(s1, s2, sequenceNum, 1, eqTerm) + tx(s1, s2, sequenceNum, 2, eqTerm);
-        } else { // Iy
+        } else { // Q
             return ty(s1, s1, sequenceNum, 0, eqTerm) + ty(s1, s1, sequenceNum, 1, eqTerm) + ty(s1, s1, sequenceNum, 2, eqTerm)
                     + ty(s1, s2, sequenceNum, 0, eqTerm) + ty(s1, s2, sequenceNum, 1, eqTerm) + ty(s1, s2, sequenceNum, 2, eqTerm);
         }
     }
 
-    public static double dIxIyij(boolean isRealPart, boolean isSide1, int sequenceNum, AsymmetricalClosedBranchCoupledCurrentEquationTerm eqTerm, Variable<AcVariableType> variable, int iDerivative) {
+    public static double dpqij(boolean isRealPart, boolean isSide1, int sequenceNum, AsymmetricalClosedBranchCoupledPowerEquationTerm eqTerm, Variable<AcVariableType> variable, int iDerivative) {
 
         int s1 = 1;
         int s2 = 2;
@@ -63,11 +65,11 @@ public class AsymmetricalClosedBranchCoupledCurrentEquationTerm extends Abstract
 
         // iDerivative is the side of "variable" that is used for derivation
         if (isRealPart) {
-            // dIx
+            // dP
             return dtx(s1, s1, sequenceNum, 0, eqTerm, variable, iDerivative) + dtx(s1, s1, sequenceNum, 1, eqTerm, variable, iDerivative) + dtx(s1, s1, sequenceNum, 2, eqTerm, variable, iDerivative)
                     + dtx(s1, s2, sequenceNum, 0, eqTerm, variable, iDerivative) + dtx(s1, s2, sequenceNum, 1, eqTerm, variable, iDerivative) + dtx(s1, s2, sequenceNum, 2, eqTerm, variable, iDerivative);
         } else {
-            // dIy
+            // dQ
             return dty(s1, s1, sequenceNum, 0, eqTerm, variable, iDerivative) + dty(s1, s1, sequenceNum, 1, eqTerm, variable, iDerivative) + dty(s1, s1, sequenceNum, 2, eqTerm, variable, iDerivative)
                     + dty(s1, s2, sequenceNum, 0, eqTerm, variable, iDerivative) + dty(s1, s2, sequenceNum, 1, eqTerm, variable, iDerivative) + dty(s1, s2, sequenceNum, 2, eqTerm, variable, iDerivative);
         }
@@ -75,18 +77,17 @@ public class AsymmetricalClosedBranchCoupledCurrentEquationTerm extends Abstract
 
     @Override
     public double eval() {
-        return ixIyij(isRealPart, isSide1, sequenceType.getNum(), this);
+        return pqij(isRealPart, isSide1, sequenceType.getNum(), this);
     }
 
     @Override
     public double der(Variable<AcVariableType> variable) {
         Objects.requireNonNull(variable);
-        return dIxIyij(isRealPart, isSide1, sequenceType.getNum(), this, variable, sideOfDerivative(variable));
+        return dpqij(isRealPart, isSide1, sequenceType.getNum(), this, variable, sideOfDerivative(variable));
     }
 
     @Override
     public String getName() {
-        return "ac_ixiy_coupled_closed_1";
+        return "ac_pq_coupled_closed_1";
     }
-
 }
